NR-384 -- Moving some classes around and adding some comments
