+++ vs/src/CodeStream.VisualStudio.Core/CodeStream.VisualStudio.Core.csproj
@@ -100,16 +100,20 @@
     <Compile Include="Logging\Sanitizer\TextProcessor.cs" />
     <Compile Include="Logging\TraceLevel.cs" />
     <Compile Include="Models\CodeStreamDiffUri.cs" />
+    <Compile Include="Models\EntityAccount.cs" />
+    <Compile Include="Models\Agent.Protocol.CodeLevelMetrics.cs" />
     <Compile Include="Models\Host.Protocol.Review.cs" />
     <Compile Include="Models\IOptions.cs" />
     <Compile Include="Models\IVirtualTextDocument.cs" />
+    <Compile Include="Models\MetricTimesliceNameMapping.cs" />
     <Compile Include="Models\NREnvironmentSettings.cs" />
     <Compile Include="Models\OpenEditorResult.cs" />
     <Compile Include="Models\Placement.cs" />
     <Compile Include="Models\Position.cs" />
+    <Compile Include="Models\RepoInfo.cs" />
     <Compile Include="Models\SessionSignedOutReason.cs" />
     <Compile Include="Models\VirtualTextDocument.cs" />
-    <Compile Include="Models\Webview.Protocol.CodeLens.cs" />
+    <Compile Include="Models\Webview.Protocol.CodeLevelMetrics.cs" />
     <Compile Include="Models\Webview.Protocol.Shell.cs" />
     <Compile Include="Packages\ToolWindowProvider.cs" />
     <Compile Include="Process\Node.cs" />
