+++ vs/src/CodeStream.VisualStudio.Core/Models/Host.Protocol.Editor.cs
@@ -119,4 +119,13 @@ namespace CodeStream.VisualStudio.Core.Models {
 		public const string MethodName = "host/marker/inserttext";
 		public override string Method => MethodName;
 	}
+
+	public class RefreshEditorsCodeLensRequestType : RequestType<EmptyRequestTypeParams> {
+		public const string MethodName = "host/editors/codelens/refresh";
+		public override string Method => MethodName;
+	}
+
+	public class RefreshEditorsCodeLensResponse {
+		public bool Success { get; set; }
+	}
 }
