+++ vs/src/CodeStream.VisualStudio.Core/Models/MetricTimesliceNameMapping.cs
@@ -1,0 +1,8 @@
+﻿namespace CodeStream.VisualStudio.Core.Models
+{
+	public class MetricTimesliceNameMapping {
+		public string D { get; set; }
+		public string T { get; set; }
+		public string E { get; set; }
+	}
+}
