+++ vs/src/CodeStream.VisualStudio.Core/Models/RepoInfo.cs
@@ -1,0 +1,8 @@
+﻿namespace CodeStream.VisualStudio.Core.Models
+{
+	public class RepoInfo {
+		public string Id { get; set; }
+		public string Name { get; set; }
+		public string Remote { get; set; }
+	}
+}
