+++ vs/src/CodeStream.VisualStudio/Services/CodeLensConnection.cs
@@ -3,6 +3,10 @@ using System.IO.Pipes;
 using StreamJsonRpc;
 
 namespace CodeStream.VisualStudio.Services {
+
+	/// <summary>
+	/// Responsible for handling the connections from Visual Studio down to the CodeLens provider
+	/// </summary>
 	public class CodeLensConnection {
 		public JsonRpc Rpc;
 		private readonly NamedPipeServerStream _stream;
